[To dev/1.3] Adjust default data region num per node from 5 to to cpu core / 2 & IoTConsensusV2: Flush old leader when leader transfer & Delay the execution of invalidateSchemaCache after leader change  (#14963)
* Adjust default data region num per node from 5 to to cpu core / 2 (#14348)

* finish


* fix it & ut


---------

(cherry picked from commit f1714d05cdb6bfd007dc1efd2c543ea5d36fcf3a)

* merge

* merge

* spotless

---------

+++ iotdb-core/confignode/src/main/java/org/apache/iotdb/confignode/manager/load/balancer/region/PartiteGraphReplicationRegionGroupAllocator.java
@@ -69,10 +69,9 @@ public class PartiteGraphReplicationRegionGroupAllocator implements IRegionGroup
       TConsensusGroupId consensusGroupId) {
 
     this.regionPerDataNode =
-        (int)
-            (consensusGroupId.getType().equals(TConsensusGroupType.DataRegion)
-                ? ConfigNodeDescriptor.getInstance().getConf().getDataRegionPerDataNode()
-                : ConfigNodeDescriptor.getInstance().getConf().getSchemaRegionPerDataNode());
+        consensusGroupId.getType().equals(TConsensusGroupType.DataRegion)
+            ? ConfigNodeDescriptor.getInstance().getConf().getDataRegionPerDataNode()
+            : ConfigNodeDescriptor.getInstance().getConf().getSchemaRegionPerDataNode();
     prepare(replicationFactor, availableDataNodeMap, allocatedRegionGroups);
 
     // Select a set of optimal alpha nodes
